+++ packages/jest-config/src/readConfigFileAndSetRootDir.ts
@@ -9,14 +9,20 @@ import * as path from 'path';
 import * as fs from 'graceful-fs';
 import parseJson = require('parse-json');
 import stripJsonComments = require('strip-json-comments');
-import type {Service} from 'ts-node';
 import type {Config} from '@jest/types';
+import {extract, parse} from 'jest-docblock';
 import {interopRequireDefault, requireOrImportModule} from 'jest-util';
 import {
   JEST_CONFIG_EXT_JSON,
   JEST_CONFIG_EXT_TS,
   PACKAGE_JSON,
 } from './constants';
+
+interface TsLoader {
+  enabled: (bool: boolean) => void;
+}
+
+type TsLoaderModule = 'ts-node' | 'esbuild-register';
 
 // Read the configuration and set its `rootDir`
 // 1. If it's a `package.json` file, we look into its "jest" property
@@ -82,7 +88,19 @@ const loadTSConfigFile = async (
   configPath: string,
 ): Promise<Config.InitialOptions> => {
   // Get registered TypeScript compiler instance
-  const registeredCompiler = await getRegisteredCompiler();
+  const docblockPragmas = parse(extract(fs.readFileSync(configPath, 'utf8')));
+  const tsLoader = docblockPragmas['jest-config-loader'] || 'ts-node';
+  if (Array.isArray(tsLoader)) {
+    throw new Error(
+      `You can only define a single test environment through docblocks, got "${tsLoader.join(
+        ', ',
+      )}"`,
+    );
+  }
+
+  const registeredCompiler = await getRegisteredCompiler(
+    tsLoader as TsLoaderModule,
+  );
 
   registeredCompiler.enabled(true);
 
@@ -98,30 +116,50 @@ const loadTSConfigFile = async (
   return configObject;
 };
 
-let registeredCompilerPromise: Promise<Service>;
+let registeredCompilerPromise: Promise<TsLoader>;
 
-function getRegisteredCompiler() {
+function getRegisteredCompiler(loader: TsLoaderModule) {
   // Cache the promise to avoid multiple registrations
-  registeredCompilerPromise = registeredCompilerPromise ?? registerTsNode();
+  registeredCompilerPromise =
+    registeredCompilerPromise ?? registerTsLoader(loader);
   return registeredCompilerPromise;
 }
 
-async function registerTsNode(): Promise<Service> {
+async function registerTsLoader(loader: TsLoaderModule): Promise<TsLoader> {
   try {
     // Register TypeScript compiler instance
-    const tsNode = await import('ts-node');
-    return tsNode.register({
-      compilerOptions: {
-        module: 'CommonJS',
-      },
-      moduleTypes: {
-        '**': 'cjs',
-      },
-    });
+    if (loader === 'ts-node') {
+      const tsLoader = await import('ts-node');
+      return tsLoader.register({
+        compilerOptions: {
+          module: 'CommonJS',
+        },
+        moduleTypes: {
+          '**': 'cjs',
+        },
+      });
+    } else if (loader === 'esbuild-register') {
+      const tsLoader = await import('esbuild-register/dist/node');
+      let instance: {unregister: () => void} | undefined;
+      return {
+        enabled: (bool: boolean) => {
+          if (bool) {
+            instance = tsLoader.register({
+              target: `node${process.version.slice(1)}`,
+            });
+          } else {
+            instance?.unregister();
+          }
+        },
+      };
+    }
+    throw new Error(
+      `Jest: '${loader}' is not a valid TypeScript configuration loader.`,
+    );
   } catch (e: any) {
     if (e.code === 'ERR_MODULE_NOT_FOUND') {
       throw new Error(
-        `Jest: 'ts-node' is required for the TypeScript configuration files. Make sure it is installed\nError: ${e.message}`,
+        `Jest: '${loader}' is required for the TypeScript configuration files. Make sure it is installed\nError: ${e.message}`,
       );
     }